added check if joined already
-prevents users from joining organization twice by hiding the view of
the join button and email field if there is a database record for that
user with the shown organization.
-issue: checking the database takes a second, so the view doesn’t
update immediately. A solution could be to add a popup form for joining
and only have a button that says “Join” or replace with an icon that
says that the user has already joined.