search results being found
-correct search results are being found and added to list in
PostSearcher class (mSearchResultsPosts).
-all potential carpool objects of search results are stored in another
list in PostSearcher class (mPotentialCarpools).
-To check if a post is possible, the total trip time is calculated and
the post will match if the driver’s departure time is after the time
that the driver needs to leave by to get to the destination on time
(meaning that all carpool participants get to the destination before
their desired arrival time).
-At the moment the algorithm doesn’t check the rider’s desired
departure time.